Las clases se renombraron para evitar problemas de compilacion
