+++ mainHire.scala
@@ -2,7 +2,7 @@
 //201124783
 //Analisis y diseño de algoritmos
 
-object main{
+object mainHire{
 	def main(args: Array[String]): Unit = {
 	    val A = Array(10,9,8,7,6,5,4,3,2,1,0)
 	    val A1 = Array(10,9,8,7,6,5,4,3,2,1,0)
